backend/requirements.txt
asgiref==3.3.4
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
